apps/plant-disease-detection/requirements.inference.txt
flask==3.1.0
numpy==2.1.3
pillow==11.0.0
tensorflow-cpu==2.20.0
